requirements.txt
pandas
requests
python-dotenv  
streamlit
matplotlib
seaborn